Updated SEB 2.2pre installer.
- Added new asymmetric/symmetric encryption for config files using
identity certificate: First encrypt a randomly generated symmetric key
using the X.509 certificate, which is then used to encrypt the config
data (for better performance when decrypting certificate encrypted
config files).
- Added new option to Config Tool / Config File pane to use old
asymmetric-only encryption (for compatibility with SEB < 2.2).
- Added new settings option in Browser pane "Allow navigating in
additional browser windows". This allows to separately allow browsing
back/forward in new browser windows (for example for additional
resources), even if it should not be allowed in the main browser window
where the exam is running.