Updated SEB 2.2pre:
- Again fixed a a bug  which sometimes caused SEB to crash while
exiting, which was caused by the Create New Desktop kiosk mode and the
therefore omitted "STAThread" command (this time properly). "STAThread"
is now active, which should increase stability of SEB while quitting
and starting (sometimes the "Confirm Quitting" message box wasn't
displayed and SEB seemed frozen when quitting).
- Fixed on-screen keyboard wasn't displayed in latest versions of
Windows 10. "Tablet Mode" needs to be activated in the system. If it
isn't, the keyboard icon in the SEB task bar is deactivated and the
tooltip indicates that "Tablet Mode" needs to be active and the
hardware keyboard detached.