Updated installer for SEB 2.2pre2:
- Enhanced SEB Windows Service implementation with fallback file for
Registry settings and added more logging information, to prevent that
lock, switch user, sign out, change a password and Task Manager are
missing in the Windows Security Screen invoked with Ctrl-Alt-Del.
- Added improved logging for the embedded browser. Note: The log path
cannot be changed in exam settings, SEB now uses always the path (and
log enable option) which is configured in client config.
- Fixed: Quitting running applications, either when starting SEB (if
they are configured as permitted applications) or when quitting SEB
(running permitted applications) failed when these applications
consisted of more than one process (have a window handling process) or
due to a bug in the quitting processes function.
- Again attempting to fix crash when quitting SEB by disabling WndProc
of TaskbarToolStrip on application shutdown.
- Removed starting seb2/Firefox with -silent argument again, as it
doesn't seem to be necessary to explicitly create seb2/Firefox profile.